[ Style ] Racing파일 추가하여 기존 index 파일의 기능 이동
